fix type for theia_active_path_timeout

+++ linux-lts-quantal-3.5.0/kernel/replay.c
@@ -133,8 +133,8 @@ int debug_flag = 0;
 #define N_SUBBUFS 4
 bool theia_active_path = 1;
 EXPORT_SYMBOL(theia_active_path);
-// stored as seconds
-unsigned int theia_active_path_timeout = 120;
+//stored as seconds
+unsigned long theia_active_path_timeout = 120;
 EXPORT_SYMBOL(theia_active_path_timeout);
 
 //#define APP_DIR		"theia_logs"
